[JENKINS-41955] - Relocate the unit test to a separate class
