+++ src/test/java/com/michelin/cio/hudson/plugins/maskpasswords/MaskPasswordsWorkflowTest.java
@@ -188,44 +188,7 @@ public class MaskPasswordsWorkflowTest {
         });
     }
     
-    @Test
-    public void passwordParameterMask() throws Exception {
-        story.addStep(new Statement() {
-            String clearTextPassword = "myClearTextPassword";
-            String logWithClearTextPassword = "printed " + clearTextPassword + " oops";
-            String logWithHiddenPassword = "printed ******** oops";
-            
-            @Override
-            public void evaluate() throws Throwable {
-                FreeStyleProject project =
-                        story.j.jenkins.createProject(FreeStyleProject.class, "testPasswordParameter");
-                
-                PasswordParameterDefinition passwordParameterDefinition =
-                        new PasswordParameterDefinition("Password1", clearTextPassword, null);
-                ParametersDefinitionProperty parametersDefinitionProperty =
-                        new ParametersDefinitionProperty(passwordParameterDefinition);
-                project.addProperty(parametersDefinitionProperty);
-                
-                MaskPasswordsBuildWrapper maskPasswordsBuildWrapper =
-                        new MaskPasswordsBuildWrapper(Collections.<MaskPasswordsBuildWrapper.VarPasswordPair>emptyList());
-                project.getBuildWrappersList().add(maskPasswordsBuildWrapper);
-                
-                project.getBuildersList().add(new TestBuilder() {
-                    @Override
-                    public boolean perform(AbstractBuild<?, ?> build, Launcher launcher, BuildListener listener) throws InterruptedException, IOException {
-                        listener.getLogger().println(logWithClearTextPassword);
-                        build.setResult(Result.SUCCESS);
-                        return true;
-                    }
-                });
-                
-                FreeStyleBuild build = project.scheduleBuild2(0, new Cause.UserIdCause()).get();
-                story.j.assertBuildStatusSuccess(story.j.waitForCompletion(build));
-                story.j.assertLogContains(logWithHiddenPassword, build);
-                story.j.assertLogNotContains(logWithClearTextPassword, build);
-            }
-        });
-    }
+
 
     // Copied from credentials-binding-plugin; perhaps belongs in JenkinsRule?
     private static Set<String> grep(File dir, String text) throws IOException {
